refactor(cli): Clean up the CLI module structure
