+++ lib/cli/src/commands/app/list.rs
@@ -1,5 +1,7 @@
+//! List Edge apps.
+
 use crate::{
-    cmd::AsyncCliCommand,
+    commands::AsyncCliCommand,
     opts::{ApiOpts, ListFormatOpts},
 };
