+++ lib/cli/src/commands/app/util.rs
