+++ lib/cli/src/commands/connect.rs
