+++ lib/cli/src/commands/namespace/create.rs
@@ -1,5 +1,5 @@
 use crate::{
-    cmd::AsyncCliCommand,
+    commands::AsyncCliCommand,
     opts::{ApiOpts, ItemFormatOpts},
 };
