+++ lib/cli/src/commands/namespace/list.rs
@@ -1,5 +1,5 @@
 use crate::{
-    cmd::AsyncCliCommand,
+    commands::AsyncCliCommand,
     opts::{ApiOpts, ListFormatOpts},
 };
